% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{initialize,JASPAR2024-method}
\alias{initialize,JASPAR2024-method}
\title{rdname JASPAR2024}
\usage{
\S4method{initialize}{JASPAR2024}(.Object, package = "JASPAR2024")
}
\description{
rdname JASPAR2024
}
