% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{JASPAR2024-class}
\alias{JASPAR2024-class}
\alias{JASPAR2024}
\title{JASPAR2024 object class}
\description{
The JASPAR2024 object class is a thin class for storing the path of
JASPAR2024 style SQLite file.
}
\section{Slots}{

\describe{
\item{\code{db}}{Object of class \code{"character"} a character string of the path
of SQLite file.}
}}

\examples{

library(JASPAR2024)
JASPAR2024

}
\author{
Damir Baranasic
}
\keyword{classes}
